Reduce duplicate console message code
